Restructured the application and deleted unimportant files
